requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose
python-decouple
python-multipart
passlib
pydantic
psycopg2-binary
pytest
playsound
shutils